requirements.txt
python-dotenv
mysql
mysql-connector-python
mysqlclient
bcrypt
cryptography